fix(ci): python-coverage-shards — split oversized overlay test part_003 under 1500-line limit
test_claude_auth_overlay_part_003.py grew to 1639 lines, tripping the
MAX_FIRST_PARTY_FILE_LINES guardrail in
test_core_decomposition_maintainability. Move the ambiguous-host /
content-compare whiteout guards, the _safe_overlay_whiteout primitive,
and the _legacy_path_confidently_absent checks into a new part_005,
importing the shared _mkdirs / _recording_mknod helpers from part_003.
No tests changed; both files are now well under the limit.
